% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.r
\name{new_stack}
\alias{new_stack}
\title{Create new stack object}
\usage{
new_stack(init = 20L)
}
\arguments{
\item{init}{an initial size of the stack}
}
\value{
an environment containing callable methods, see details
}
\description{
Create a new stack object with reference semantics and two methods,
`push` adds objects to the stack and `pop` removes them.
}
\details{
This is dependency-free class based on function closures with reference semantics.
Internally, the stack is represented as a pre-allocated `list` that is extended as required.
The size of the `list` is set to the `init`, and is dynamically extended by the `init` value.

Following methods are available:

* `push(...)` - element or elements to stack, elements are converted to list
* `pop(n)` - removes and returns `n` latest elements from stack, if `n` is missing, remove all elements

Following slots are available:

* items - stack memory, a pre-allocated
* size - current size of the stack

Modifying these slots could lead to inconsistent behaviour.
}
\examples{
s = new_stack()
s$push("foo", "bar", "baz")

identical(s$pop(1), list("foo"))
identical(s$pop(), list("bar", "baz"))
identical(s$pop(), list())

}
