% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutr.r
\name{test}
\alias{test}
\title{Test expression}
\usage{
test(expr)
}
\arguments{
\item{expr}{an expression to be evaluated}
}
\value{
an evaluated expression
}
\description{
Test provided expression for true-ness, the evaluated expression is returned invisibly.
If the `.mutr` object exists, the test counter is increased and the error message is appended.
Otherwise, the error message is printed.
}
\details{
Any other condition than `TRUE` results in a differe with different diagnostic expression
depending on the evaluation. For this purpose, `test` distinguishes between:

* error condition
* non-logical object
* logical object with length > 1
* singular logical FALSE
* singular logical TRUE
}
\examples{
test(TRUE)
test(FALSE)
test(stop())
}
