% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.r
\name{test_pkg}
\alias{test_pkg}
\alias{check}
\alias{document}
\title{Package development functions}
\usage{
test_pkg(pkg = ".")

check(pkg = ".", as_cran = FALSE)

document(pkg = ".")
}
\arguments{
\item{pkg}{a path to the package directory}

\item{as_cran}{run the `R CMD check` with the CRAN preset (internet connection required)}
}
\value{
nothing, these functions are run for their side-effect
}
\description{
These functions provide a dependency-light alternative to `devtools`.
}
\details{
`test_pkg()` install and runs tests for specified package, these tests must be created by
some unit-testing framework in that they need to be able to be run outside of `R CMD check`.

The `check()` maps to the `R CMD check` and `document()` runs the `roxygen2::roxygenize`
function to generate documentation.
}
